+++ packages/adminui-vue/components/UTableEntity/filter-templates/entity/isNull.vue
@@ -13,10 +13,11 @@ export default {
   mixins: [require('../mixinForFilter.js')],
 
   methods: {
-    getCondition() {
+    getCondition () {
+      const { $ut } = this
       return {
         whereList: [{ condition: 'isNull' }],
-        description: 'isNull'
+        description: $ut('isNull')
       }
     }
   }