+++ packages/adminui-vue/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+- filters `isNull` and `notIsNull` - localised description
  - `UToolbar`: enable dropdown if there are slots `dropdown` or `dropdown-prepend`
 
 ## [5.23.19] - 2022-09-12